Repurpose diff-flags parameter and use 3-way merge in diff
* `diff-flags` is now used to define additional flags for `helm diff
  upgrade` on top of the ones passed by `upgrade-flags`. Unknown
  parameters are ignored by the diff plugin now.
* `--three-way-merge` is now the default value for `diff-flags`, making
  the diff perform a 3-way merge to align with the behaviour of `helm
  upgrade`.
* Use `--three-way-merge` in install script.

Closes #569.
Closes #574.
