+++ docs/tasks/ods-deploy-helm.adoc
@@ -80,13 +80,13 @@ The following artifacts are generated by the task and placed into `.ods/artifact
 
 
 | diff-flags
-| --install
-| Flags to pass to `helm diff upgrade`. Note that `--detailed-exitcode` and `--no-color` are automatically set and cannot be removed. Changes should be aligned with `upgrade-flags` as needed.
+| --three-way-merge
+| Flags to pass to `helm diff upgrade` in addition to the ones specified via the `upgrade-flags` parameter. Note that the flags `--detailed-exitcode` and `--no-color` are automatically set and cannot be removed. If flags unknown to `helm diff` are passed, they are ignored.
 
 
 | upgrade-flags
 | --install --wait
-| Flags to pass to `helm upgrade`. Changes should be aligned with `diff-flags` as needed.
+| Flags to pass to `helm upgrade`.
 
 
 | age-key-secret
